Relocate DB Related Files
