+++ compose_api/common/gateway/utils.py
@@ -1,4 +1,4 @@
-from compose_api.simulation.database_service import DatabaseService
+from compose_api.db.database_service import DatabaseService
 from compose_api.simulation.models import HpcRun, JobType
