+++ compose_api/simulation/simulation_service.py
@@ -18,7 +18,7 @@ from compose_api.common.hpc.models import SlurmJob
 from compose_api.common.hpc.slurm_service import SlurmService
 from compose_api.common.ssh.ssh_service import SSHService
 from compose_api.config import get_settings
-from compose_api.simulation.database_service import DatabaseService
+from compose_api.db.database_service import DatabaseService
 from compose_api.simulation.hpc_utils import (
     get_slurm_log_file,
     get_slurm_sim_experiment_dir,
